+++ packages/orq-rc/src/models/operations/updatedataset.ts
@@ -182,7 +182,7 @@ export const UpdateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-07-21T01:16:37.138Z",
+    "2026-07-21T08:06:01.267Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {