+++ packages/orq-rc/RELEASES.md
@@ -10979,3 +10979,13 @@ Based on:
 - [typescript v4.12.0-rc.44] packages/orq-rc
 ### Releases
 - [NPM v4.12.0-rc.44] https://www.npmjs.com/package/@orq-ai/node/v/4.12.0-rc.44 - packages/orq-rc
+
+## 2026-07-21 08:09:02
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.790.3 (2.918.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.12.0-rc.45] packages/orq-rc
+### Releases
+- [NPM v4.12.0-rc.45] https://www.npmjs.com/package/@orq-ai/node/v/4.12.0-rc.45 - packages/orq-rc